requirements.txt
Flask==1.1.2
SQLAlchemy==1.3.20
Flask-WTF==0.14.3
Pillow==8.0.1
